requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
zcrmsdk
sendgrid==6.11.0
apscheduler==3.10.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0